processing_service/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
psycopg2==2.9.9
requests==2.31.0
